# Make (de)serializer more flexible