--- a/libcaf_core/CMakeLists.txt
+++ b/libcaf_core/CMakeLists.txt
@@ -33,8 +33,6 @@ set (LIBCAF_CORE_SRCS
      src/behavior.cpp
      src/behavior_stack.cpp
      src/behavior_impl.cpp
-     src/binary_deserializer.cpp
-     src/binary_serializer.cpp
      src/blocking_actor.cpp
      src/concatenated_tuple.cpp
      src/continue_helper.cpp