--- a/libcaf_core/caf/deserializer.hpp
+++ b/libcaf_core/caf/deserializer.hpp
@@ -42,7 +42,7 @@ public:
 
   explicit deserializer(actor_system& sys);
 
-  explicit deserializer(execution_unit* ctx);
+  explicit deserializer(execution_unit* ctx = nullptr);
 };
 
 /// Reads `x` from `source`.