… libcaf_core/caf/fwd.hpp
… libcaf_core/caf/streambuf.hpp
… libcaf_core/src/binary_deserializer.cpp
… libcaf_core/src/binary_serializer.cpp
… libcaf_core/test/streambuf.cpp