First, this change moves the context to the parameter to the end in order to make it optional. The context is only necessary when serializing messages. Since CAF's serialization framework also serves also other application's need, this CAF-specific detail now no longer constitutes a mandatory part of the API.

This change also unifies the constrution of caf::serializer from a function that writes a pair of const char* pointer and size into an arbirtrary destination.

Status:

- [x] Make CAF context parameter optional
- [x] Add new streambuffer types
- [x] Make serializer API streambuffer-based
- [x] Make deserializer API streambuffer-based
- [x] Perform rough viability performance comparison
- [x] Perform end-to-end performance comparison