@mavam 2016-04-14
Here's the more comprehensive benchmark:

### Infrastructure

```cpp
#include <caf/all.hpp>

using namespace std;
using namespace std::chrono;
using namespace caf;

using stopwatch = high_resolution_clock;

// Measure overhead of data serialization performance.
template <class T>
microseconds benchmark(T x) {
  vector<char> buffer;
  // Serialize.
#ifdef NEW
  vectorbuf<char> vb{buffer};
  binary_serializer bs{vb};
#else
  scoped_execution_unit ctx;
  binary_serializer bs{&ctx, buffer};
#endif
  auto start = stopwatch::now();
  for (auto i = 0u; i < 100000; ++i)
    bs << x;
  // Deserialize.
#ifdef NEW
  charbuf cb{buffer};
  binary_deserializer bd{cb};
#else
  binary_deserializer bd{&ctx, buffer};
#endif
  T y;
  for (auto i = 0u; i < 100000; ++i)
    bd >> y;
  auto stop = stopwatch::now();
  return duration_cast<microseconds>(stop - start);
}

// Measure overhead of construction, since serializers are often used in a
// ad-hoc fashion.
template <class T>
microseconds benchmark2(T x) {
#ifndef NEW
  scoped_execution_unit ctx;
#endif
  auto start = stopwatch::now();
  for (auto i = 0u; i < 100000; ++i) {
    vector<char> buffer;
#ifdef NEW
    vectorbuf<char> vb{buffer};
    binary_serializer bs{vb};
#else
    binary_serializer bs{&ctx, buffer};
#endif
    bs << x;
  // Deserialize.
#ifdef NEW
    charbuf cb{buffer};
    binary_deserializer bd{cb};
#else
    binary_deserializer bd{&ctx, buffer};
#endif
    T y;
    bd >> y;
  }
  auto stop = stopwatch::now();
  return duration_cast<microseconds>(stop - start);
}

auto main() -> int {
  auto str = std::string{"the quick brown fox jumps over the lazy dog"};
  auto large = str;
  for (auto i = 0; i < 100; ++i)
    large += str;
  auto u64 = uint64_t{42};
  cout << benchmark(str).count() << "us" << endl;
  cout << benchmark(large).count() << "us" << endl;
  cout << benchmark(u64).count() << "us" << endl;
  cout << "-------" << endl;
  cout << benchmark2(str).count() << "us" << endl;
  cout << benchmark2(large).count() << "us" << endl;
  cout << benchmark2(u64).count() << "us" << endl;
}
```

System: Mac OS X (Yosemite)
Compiler: Clang "7.02"

Compiled as follows:
- `c++ -O3 -g -std=c++11 -stdlib=libc++ -I ~/code/caf/libcaf_core -L ~/code/caf/build/lib -DNEW benchmark.cpp -o bench-new -lcaf_core`
- `c++ -O3 -g -std=c++11 -stdlib=libc++ -I ~/code/caf/libcaf_core -L ~/code/caf/build-as/lib benchmark.cpp -o bench-old -lcaf_core`

### Results

Old:
```
32811us
2755070us
2822us
-------
51900us
1964340us
10060us
```

NEW:
```
17125us
1077772us
2891us
-------
90098us
877076us
32766us
```

Multiple runs give numbers in the same range.

Difference Benchmark 1:
- 1.92
- 2.56
- 0.98

Difference Benchmark 2:
- 0.58
- 2.24
- 0.31

The difference represents the fraction of *old* divided by *new*, i.e., the relative change to the existing code base.

### Interpretation

The new change increases the throughput up to a factor of 2.5, at a slight setup cost. As soon as the system pumps a reasonable amount of larger data through the pipeline, this overhead is quickly amortized.

However, the "construction overhead" represents a worst-case scenario. A real-world scenario does not construct a serializer/deserializer in a loop of 100K iterations.

Overall, I conclude that the huge performance gains of abstraction and throughput outweigh the somewhat more expensive setup costs.