@Neverlord 2016-04-18
I think the new interface is quite clumsy and repetitive on the caller side. I would prefer an approach where the caller side doesn't change. Instead of inverting the order on the ctor, you could as well introduce a new constructor that takes a single argument. It's one extra ctor, but it keeps the interface stable.

Since CAF will always use the same buffer types with `binary_serializer` and `binary_deserializer`, I would prefer having the buffer simply as member (again, to keep the interface as-is). This would allow us to keep the caller-side stable and get around the tedious "make a buffer from the vector, then initialize the (de)serializer".

You could template the current `binary_serializer` implementation as `stream_serializer<T>` to configure what member it has and then `using binary_serializer = stream_serializer<vectorbuf<char>>`. Users that wish to store the streambuf separately can simply have `using my_serializer = stream_serializer<std::streambuf&>` instead.

Looking at your (pretty good!) evaluation, I think we should consider storing the serializers as members in the long run instead of the current ad-hoc approach to safe us the ctor/dtor overhead.