@mavam 2016-04-18
> I think the new interface is quite clumsy and repetitive on the caller side.

Could you elaborate on the clumsiness? Is it the 2-step initialization? I find it actually more elegant because it decouples the source/sink aspect from the rest of the serializer initialization. In my opinion, the previously hard-coded ctors made the serializers very inflexible. Additionally, Boost.Serialization's binary serializers also take a `std::streambuf` as ctor argument, which seemed to be an important consideration in the overall framework design. I do agree that it's more code and thus a bit more verbose, if that's what you mean. Conversely, it's more explicit. Since this is targeting 0.15, it seemed a reasonable design choice to me.

> Since CAF will always use the same buffer types with `binary_serializer` and `binary_deserializer`, I would prefer having the buffer simply as member (again, to keep the interface as-is). This would allow us to keep the caller-side stable and get around the tedious "make a buffer from the vector, then initialize the (de)serializer".

I was designing the interface such that other applications (in particular, VAST and Broker) can directly using the means provided by CAF. I have various other streambuffers that I'd like to plug in the binary serializers. Hence the decoupling.

>You could template the current binary_serializer implementation as stream_serializer<T> to configure what member it has and then using binary_serializer = stream_serializer<vectorbuf<char>>. Users that wish to store the streambuf separately can simply have using my_serializer = stream_serializer<std::streambuf&> instead.

I'm a bit confused here. First, I'm surprised that *you* are advocating for a template-based design where it's actually not necessary :wink:. Second, `std::streambuf` is a virtual class, so the serializer only needs a reference to the base class. But since the instantiation paths are known at compile-time, the compiler can apply devirtualization optimizations. Consequently, we can keep the serializers generic without resorting to templates.

> [..] I think we should consider storing the serializers as members in the long run instead of the current ad-hoc approach to safe us the ctor/dtor overhead.

Agreed.

Another aspect: I am still thinking about a `char` version for `vectorbuf<char>`. It would make sense to have one such that CAF's common case of `vector<char>` serialization is symmetric. Currently, there's an `arraybuf<T>` with an alias called `charbuf = arraybuf<char>`. I want the same for the streambuffer `vectorbuf<T>` which has an unbounded put area. Some suggestions:

- `vectorbuf = containerbuf<std::vector<char>>>
- `uvectorbuf = containerbuf<std::vector<uint8_t>>>` (not necessarily in CAF)

Or just `bufferbuf = containerbuf<std::vector<char>>>`, since I see `buffer = std::vector<char>` in the BASP implementation.